Solutions/ai-bartender-agent/aws/services/api/src/deleteDrink/requirements.txt
aws-lambda-powertools[tracer]>=2.0.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
